create menu component
