+++ src/_styles/_base.scss
@@ -124,3 +124,11 @@ input {
   -webkit-appearance: none;
   border-radius: 0;
 }
+
+button {
+  padding: 0;
+  border: none;
+  font: inherit;
+  color: inherit;
+  background-color: transparent;
+}
