+++ src/_styles/index.scss
@@ -4,3 +4,4 @@
 @import "layout";
 @import "fonts";
 @import "navigation";
+@import "menu";
